% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheScribe.R
\name{classify_hematopoietic}
\alias{classify_hematopoietic}
\title{Classify hematopoietic vs non-hematopoietic cells}
\usage{
classify_hematopoietic(query, ref, return.seurat = FALSE)
}
\arguments{
\item{query}{Seurat object to annotate}

\item{ref}{Seurat reference object containing PCA reduction and `nonhematopoietic` metadata}

\item{return.seurat}{Logical. If TRUE, returns Seurat object with predictions added; otherwise, return prediction table. (Default: FALSE)}
}
\value{
Either Seurat object or TransferData output
}
\description{
A prefiltering step to remove hematopoietic cells based on an integrated
dataset R1 consisting of whole bone marrow, hematopoietic stem and progenitor
cells, and central marrow and endosteal stromal cells.
}
\details{
Uses Seurat label transfer (FindTransferAnchors + TransferData)
based on PCA space from the reference.
}
