% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheScribe.R
\name{classify_stroma}
\alias{classify_stroma}
\title{Classify stroma cell types}
\usage{
classify_stroma(query, ref, return.seurat = FALSE)
}
\arguments{
\item{query}{Seurat object to annotate (typically stromal subset)}

\item{ref}{Stromal reference Seurat object containing PCA reduction and `nonhematopoietic` metadata}

\item{return.seurat}{Logical. If TRUE, returns Seurat object with predictions added; otherwise, return prediction table. (Default: FALSE)}
}
\value{
Either Seurat object or list of TransferData output
}
\description{
Performs non-hematopoietic cell annotation based on two references.
Reference R2.1 is generated by barcoding 10 FACS-purified populations of
stromal cells using on-chip multiplexing before sequencing. Reference R2.2 is
generated by curating key marker genes for stromal cell types.
}
\details{
Uses Seurat label transfer (FindTransferAnchors + TransferData)
based on PCA space from the reference. Transfers two sets of annotations: hash_label and manual_annot.
}
