% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_reference}
\alias{download_reference}
\title{Download reference data}
\usage{
download_reference(dataDir = NA)
}
\arguments{
\item{dataDir}{Specifies where the reference data should be downloaded. Should essentially never be altered by the user directly since the package looks for the data in a specific location.}
}
\description{
Downloads the reference data from the GitHub repo.
}
