% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{InflammationPlot}
\alias{InflammationPlot}
\title{NicheScribe iMSC-L score plotting function}
\usage{
InflammationPlot(query, group = NULL, ...)
}
\arguments{
\item{query}{Seurat object containing a nichescribe_umap reduction and
iMSCL_score metadata column.}

\item{group}{Optional metadata column to split the plot by.}

\item{...}{Additional arguments passed to Seurat::FeaturePlot().}
}
\value{
A ggplot or patchwork object.
}
\description{
Plots the inflammatory MSC-L score on the NicheScribe UMAP.
Optionally splits the plot by a metadata variable.
}
