% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{NichePlot}
\alias{NichePlot}
\title{NicheScribe UMAP plotting function}
\usage{
NichePlot(query, layer = c("R2.1", "R2.2"), ...)
}
\arguments{
\item{query}{Seurat object to plot from - must have NicheScribe annotations and UMAP}

\item{layer}{Annotation layer to plot. Either `"R2.1"` (uses `hash_label_pred`) or `"R2.2"` (uses `manual_annot_pred`).}

\item{...}{Additional arguments passed to Seurat::DimPlot().}
}
\value{
Seurat DimPlot using NicheScribe UMAP reduction, selected annotations, and standard color palette
}
\description{
Plots cells on NicheScribe-generated UMAP with NicheScribe annotations and standard colors
}
