% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheScribe.R
\name{score_imscl_ucell}
\alias{score_imscl_ucell}
\title{iMSC-L scoring}
\usage{
score_imscl_ucell(query, genes)
}
\arguments{
\item{query}{Seurat object to score (typically MSC-Ls).}

\item{genes}{List of 100 genes for iMSC_L score}
}
\value{
Matrix with UCell scores.
}
\description{
Computes a iMSC-L score based on a list of 100 genes upregulated in
pIC-injected MSC-L, CM_MSCL, Endo_MSCL, CM_iMSCL, and Endo_iMSCL using UCell.
}
