% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_reference}
\alias{download_reference}
\title{Download NicheScribe reference data}
\usage{
download_reference(force = FALSE)
}
\arguments{
\item{force}{Logical. If `TRUE`, forces a re-download even if the data is already up-to-date and validated. Defaults to `FALSE`.}
}
\value{
Invisible `TRUE` if successful.
}
\description{
Downloads and installs the required reference datasets for
NicheScribe from GitHub. File integrity is verified via SHA256 checksums.
}
